requirements.txt
psycopg2-binary
reportlab
Pillow
python-dotenv
